PUBDEV-6271 - IRF Mojo path length is of type integer and should be long

+++ h2o-genmodel/src/main/java/hex/genmodel/algos/isofor/IsolationForestMojoModel.java
@@ -5,8 +5,8 @@ import hex.genmodel.algos.tree.SharedTreeMojoModel;
 
 public final class IsolationForestMojoModel extends SharedTreeMojoModel {
 
-  int _min_path_length;
-  int _max_path_length;
+  long _min_path_length;
+  long _max_path_length;
 
   public IsolationForestMojoModel(String[] columns, String[][] domains, String responseColumn) {
     super(columns, domains, responseColumn);
